refactor(purchase-orders): decompose into Query/Lifecycle/Receiving services
Behavior-preserving (write-heavy, no JE). Facade keeps 16-method surface + 1-arg
ctor + delegates; internally constructs 3 sub-services (specs + module untouched).
4 $transaction (2 Serializable) moved WHOLE into single sub-service methods;
duplicated product-name builder extracted to a pure util. tsc 0, tests green, lint 0.
